+++ main/csi_data.h
@@ -11,7 +11,7 @@
  */
 typedef struct csi_data {
   uint8_t device_id; /**< Device ID that has sent the CSI data (Tx) */
-  uint16_t time_index; /**< Time index when the CSI data was sent/received */
+  uint32_t time_index; /**< Time index when the CSI data was sent/received */
   wifi_pkt_rx_ctrl_t rx_ctrl; /**< RX control information */
   uint8_t mac[6]; /**< Source MAC address of the CSI data */
   uint8_t dmac[6]; /**< Destination MAC address of the CSI data */
@@ -26,7 +26,7 @@ typedef struct csi_data {
  */
 typedef struct payload {
   uint8_t device_id; /**< Device ID that has received the CSI data (Rx) */
-  uint16_t time_index; /**< Time index when the payload was sent/received */
+  uint32_t time_index; /**< Time index when the payload was sent/received */
   uint8_t csi_data_arr_len; /**< Number of CSI data entries */
   csi_data_t csi_data_arr[CSI_DATA_ARR_LEN]; /**< Array of CSI data */
 } payload_t;
@@ -44,7 +44,7 @@ extern uint8_t last_id;
 /**
  * The time index of the last received CSI data.
  */
-extern uint16_t last_time_index;
+extern uint32_t last_time_index;
 
 /**
  * The number of loop iterations before sending CSI data.
@@ -62,7 +62,7 @@ extern uint8_t payload_index;
  *
  * @param time_index The time index of the payload
  */
-void send_csi_data(uint16_t time_index);
+void send_csi_data(uint32_t time_index);
 
 /**
  * @brief Callback function to handle received CSI data.
